requirements.txt
anyio==3.6.2
deta==1.1.0
fastapi==0.88.0
idna==3.4
pydantic==1.10.2
python-dotenv==0.21.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
